requirements.txt
# Core data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# OpenAI API
openai>=0.28.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Additional utilities
requests>=2.31.0
beautifulsoup4>=4.12.0

# Logging and configuration
python-dotenv>=1.0.0
